chore: unify benchmark scripts into bench.py
Replaces codegen-lines.py, codegen-lines.sh, jump-resolution.py,
jump-resolution.sh, and const-input-stats.py with a single bench.py.

Flags: --codegen (default), --jump-resolution, --input-stats, --diff REV,
--extra-dir, --no-codegen-lines, --no-compile-time.
